:art: #5 ソーシャルログインの場合、サインアップとサインインで処理を分ける必要はないのでエンドポイントを統一した
