+++ src/app/google/_components/GoogleLoginButton/index.ts
@@ -1,0 +1,1 @@
+export { GoogleLoginButton } from './GoogleLoginButton';
