+++ src/app/google/_components/index.ts
@@ -1,0 +1,1 @@
+export { GoogleLoginButton } from './GoogleLoginButton';
