+++ src/app/page.tsx
@@ -47,6 +47,8 @@ export default async function Index() {
     data: { user },
   } = await supabase.auth.getUser();
 
+  console.log(user);
+
   return (
     <div className="w-full flex flex-col items-center">
       <nav className="w-full flex justify-center border-b border-b-foreground/10 h-16">
@@ -67,16 +69,10 @@ export default async function Index() {
                   Login
                 </Link>
                 <Link
-                  href="/google/signup"
+                  href="/google/login"
                   className="py-2 px-4 rounded-md no-underline bg-btn-background hover:bg-btn-background-hover"
                 >
-                  google sign up
-                </Link>
-                <Link
-                  href="/google/signin"
-                  className="py-2 px-4 rounded-md no-underline bg-btn-background hover:bg-btn-background-hover"
-                >
-                  google sign in
+                  Googleでログイン
                 </Link>
               </div>
             )}
